backend/requirements.txt
fastapi==0.128.0
uvicorn[standard]==0.40.0
pydantic==2.12.5
pydantic-settings==2.12.0
python-multipart==0.0.21
bcrypt==5.0.0
PyJWT==2.8.0

aiomysql==0.3.2
redis[asyncio]==7.1.0
qdrant-client==1.16.2

celery==5.6.2
cos-python-sdk-v5==1.9.25

numpy==2.4.1
pandas==2.3.3
polars==1.7.0
Pillow==12.1.0

# AI 图片分析依赖
torch>=2.9.0
transformers>=4.40.0
accelerate>=0.30.0

aiohttp>=3.11.11
anyio==4.6.2

python-dotenv==1.2.1
openpyxl==3.1.5
